chore: move FpjsProvider to Providers
