+++ src/pages/_app.tsx
@@ -1,17 +1,8 @@
 import '../styles/global-styles.scss';
 import Head from 'next/head';
-import { FpjsProvider, FingerprintJSPro } from '@fingerprintjs/fingerprintjs-pro-react';
 import { AppProps } from 'next/app';
 import Providers from '../Providers';
 import { Layout } from '../Layout';
-import { env } from '../env';
-
-export const FP_LOAD_OPTIONS: FingerprintJSPro.LoadOptions = {
-  apiKey: env.NEXT_PUBLIC_API_KEY,
-  scriptUrlPattern: [env.NEXT_PUBLIC_SCRIPT_URL_PATTERN, FingerprintJSPro.defaultScriptUrlPattern],
-  endpoint: [env.NEXT_PUBLIC_ENDPOINT, FingerprintJSPro.defaultEndpoint],
-  region: env.NEXT_PUBLIC_REGION,
-};
 
 export type CustomPageProps = { embed?: boolean };
 
@@ -24,11 +15,9 @@ function CustomApp({ Component, pageProps }: AppProps<CustomPageProps>) {
         <title>Fingerprint Pro Use Cases</title>
       </Head>
       <Providers>
-        <FpjsProvider loadOptions={FP_LOAD_OPTIONS}>
-          <Layout embed={Boolean(pageProps.embed)}>
-            <Component {...pageProps} />
-          </Layout>
-        </FpjsProvider>
+        <Layout embed={Boolean(pageProps.embed)}>
+          <Component {...pageProps} />
+        </Layout>
       </Providers>
     </>
   );
